Move net::cookie unit tests into the unit_tests crate.
