+++ tests/unit/lib.rs
@@ -1,6 +1,10 @@
 extern crate geom;
 extern crate gfx;
+extern crate net;
+extern crate net_traits;
 extern crate util;
 
-#[cfg(test)] #[path="util/mod.rs"] mod util_tests;
 #[cfg(test)] #[path="gfx/mod.rs"] mod gfx_tests;
+#[cfg(test)] #[path="net/mod.rs"] mod net_tests;
+#[cfg(test)] #[path="util/mod.rs"] mod util_tests;
+
