+++ tests/unit/net/mod.rs
@@ -1,0 +1,1 @@
+mod cookie;
